adapted to new WinSso version Fixed test
