+++ src/proxy/interfaces/i.win-sso.facade.ts
